+++ dumper/mock/mock.go
@@ -1,0 +1,16 @@
+package mock
+
+import (
+	"database/sql"
+	"testing"
+
+	"github.com/DATA-DOG/go-sqlmock"
+	"github.com/stretchr/testify/assert"
+)
+
+// GetDB which can be used for testing purposes.
+func GetDB(t *testing.T) (*sql.DB, sqlmock.Sqlmock) {
+	db, mock, err := sqlmock.New()
+	assert.Nil(t, err)
+	return db, mock
+}
