+++ dumper/mysql_test.go
@@ -2,31 +2,16 @@ package dumper
 
 import (
 	"bytes"
-	"database/sql"
 	"errors"
-	"strings"
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
+	"github.com/hgfischer/mysqlsuperdump/dumper/mock"
 	"github.com/stretchr/testify/assert"
 )
 
-func getDB(t *testing.T) (*sql.DB, sqlmock.Sqlmock) {
-	db, mock, err := sqlmock.New()
-	assert.Nil(t, err)
-	return db, mock
-}
-
-func TestMySQLLockTableRead(t *testing.T) {
-	db, mock := getDB(t)
-	dumper := NewMySQLDumper(db)
-	mock.ExpectExec("LOCK TABLES `table` READ").WillReturnResult(sqlmock.NewResult(0, 1))
-	_, err := dumper.LockTableReading("table")
-	assert.Nil(t, err)
-}
-
 func TestMySQLFlushTable(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectExec("FLUSH TABLES `table`").WillReturnResult(sqlmock.NewResult(0, 1))
 	_, err := dumper.FlushTable("table")
@@ -34,42 +19,50 @@ func TestMySQLFlushTable(t *testing.T) {
 }
 
 func TestMySQLUnlockTables(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectExec("UNLOCK TABLES").WillReturnResult(sqlmock.NewResult(0, 1))
 	_, err := dumper.UnlockTables()
 	assert.Nil(t, err)
 }
 
-func TestMySQLGetTables(t *testing.T) {
-	db, mock := getDB(t)
+func TestMySQLQueryTables(t *testing.T) {
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnRows(
 		sqlmock.NewRows([]string{"Tables_in_database", "Table_type"}).
 			AddRow("table1", "BASE TABLE").
 			AddRow("table2", "BASE TABLE"),
 	)
-	tables, err := dumper.GetTables()
+	tables, err := dumper.QueryTables()
 	assert.Equal(t, []string{"table1", "table2"}, tables)
 	assert.Nil(t, err)
 }
 
+func TestMySQLLockTableRead(t *testing.T) {
+	db, mock := mock.GetDB(t)
+	dumper := NewMySQLDumper(db)
+	mock.ExpectExec("LOCK TABLES `table` READ").WillReturnResult(sqlmock.NewResult(0, 1))
+	_, err := dumper.LockTableReading("table")
+	assert.Nil(t, err)
+}
+
 func TestMySQLGetTablesHandlingErrorWhenListingTables(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	expectedErr := errors.New("broken")
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnError(expectedErr)
-	tables, err := dumper.GetTables()
+	tables, err := dumper.QueryTables()
 	assert.Equal(t, []string{}, tables)
 	assert.Equal(t, expectedErr, err)
 }
 
 func TestMySQLGetTablesHandlingErrorWhenScanningRow(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectQuery("SHOW FULL TABLES").WillReturnRows(
 		sqlmock.NewRows([]string{"Tables_in_database", "Table_type"}).AddRow(1, nil))
-	tables, err := dumper.GetTables()
+	tables, err := dumper.QueryTables()
 	assert.Equal(t, []string{}, tables)
 	assert.NotNil(t, err)
 }
@@ -80,7 +73,7 @@ func TestMySQLDumpCreateTable(t *testing.T) {
 		"`name` varchar(255) NOT NULL, " +
 		"PRIMARY KEY (`id`), KEY `idx_name` (`name`) " +
 		") ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8"
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectQuery("SHOW CREATE TABLE `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"Table", "Create Table"}).
@@ -93,7 +86,7 @@ func TestMySQLDumpCreateTable(t *testing.T) {
 }
 
 func TestMySQLDumpCreateTableHandlingErrorWhenScanningRows(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	mock.ExpectQuery("SHOW CREATE TABLE `table`").WillReturnRows(
 		sqlmock.NewRows([]string{"Table", "Create Table"}).AddRow("table", nil))
@@ -102,29 +95,29 @@ func TestMySQLDumpCreateTableHandlingErrorWhenScanningRows(t *testing.T) {
 }
 
 func TestMySQLGetColumnsForSelect(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.SelectMap = map[string]map[string]string{"table": {"col2": "NOW()"}}
 	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
 		sqlmock.NewRows([]string{"col1", "col2", "col3"}).AddRow("a", "b", "c"))
-	columns, err := dumper.GetColumnsForSelect("table")
+	columns, err := dumper.QueryColumnsForTable("table")
 	assert.Nil(t, err)
 	assert.Equal(t, []string{"`col1`", "NOW() AS `col2`", "`col3`"}, columns)
 }
 
 func TestMySQLGetColumnsForSelectHandlingErrorWhenQuerying(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.SelectMap = map[string]map[string]string{"table": {"col2": "NOW()"}}
 	error := errors.New("broken")
 	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnError(error)
-	columns, err := dumper.GetColumnsForSelect("table")
+	columns, err := dumper.QueryColumnsForTable("table")
 	assert.Equal(t, err, error)
 	assert.Empty(t, columns)
 }
 
 func TestMySQLGetSelectQueryFor(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.SelectMap = map[string]map[string]string{"table": {"c2": "NOW()"}}
 	dumper.WhereMap = map[string]string{"table": "c1 > 0"}
@@ -136,7 +129,7 @@ func TestMySQLGetSelectQueryFor(t *testing.T) {
 }
 
 func TestMySQLGetSelectQueryForHandlingError(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.SelectMap = map[string]map[string]string{"table": {"c2": "NOW()"}}
 	dumper.WhereMap = map[string]string{"table": "c1 > 0"}
@@ -148,7 +141,7 @@ func TestMySQLGetSelectQueryForHandlingError(t *testing.T) {
 }
 
 func TestMySQLGetRowCount(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.WhereMap = map[string]string{"table": "c1 > 0"}
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
@@ -159,7 +152,7 @@ func TestMySQLGetRowCount(t *testing.T) {
 }
 
 func TestMySQLGetRowCountHandlingError(t *testing.T) {
-	db, mock := getDB(t)
+	db, mock := mock.GetDB(t)
 	dumper := NewMySQLDumper(db)
 	dumper.WhereMap = map[string]string{"table": "c1 > 0"}
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
@@ -167,81 +160,4 @@ func TestMySQLGetRowCountHandlingError(t *testing.T) {
 	count, err := dumper.GetRowCountForTable("table")
 	assert.NotNil(t, err)
 	assert.Equal(t, uint64(0), count)
-}
-
-func TestMySQLDumpTableHeader(t *testing.T) {
-	db, mock := getDB(t)
-	dumper := NewMySQLDumper(db)
-	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table`").WillReturnRows(
-		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(1234))
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	count, err := dumper.WriteTableHeader(buffer, "table")
-	assert.Equal(t, uint64(1234), count)
-	assert.Nil(t, err)
-	assert.Contains(t, buffer.String(), "Data for table `table`")
-	assert.Contains(t, buffer.String(), "1234 rows")
-}
-
-func TestMySQLDumpTableHeaderHandlingError(t *testing.T) {
-	db, mock := getDB(t)
-	dumper := NewMySQLDumper(db)
-	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table`").WillReturnRows(
-		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(nil))
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	count, err := dumper.WriteTableHeader(buffer, "table")
-	assert.Equal(t, uint64(0), count)
-	assert.NotNil(t, err)
-}
-
-func TestMySQLDumpTableLockWrite(t *testing.T) {
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewMySQLDumper(nil)
-	dumper.WriteTableLockWrite(buffer, "table")
-	assert.Contains(t, buffer.String(), "LOCK TABLES `table` WRITE;")
-}
-
-func TestMySQLDumpUnlockTables(t *testing.T) {
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewMySQLDumper(nil)
-	dumper.WriteUnlockTables(buffer)
-	assert.Contains(t, buffer.String(), "UNLOCK TABLES;")
-}
-
-func TestMySQLDumpTableData(t *testing.T) {
-	db, mock := getDB(t)
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewMySQLDumper(db)
-	dumper.ExtendedInsertRows = 2
-
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
-		sqlmock.NewRows([]string{"id", "language"}).
-			AddRow(1, "Go"))
-
-	mock.ExpectQuery("SELECT `id`, `language` FROM `table`").WillReturnRows(
-		sqlmock.NewRows([]string{"id", "language"}).
-			AddRow(1, "Go").
-			AddRow(2, "Java").
-			AddRow(3, "C").
-			AddRow(4, "C++").
-			AddRow(5, "Rust").
-			AddRow(6, "Closure"))
-
-	assert.Nil(t, dumper.WriteTableData(buffer, "table"))
-
-	assert.Equal(t, strings.Count(buffer.String(), "INSERT INTO `table` VALUES"), 3)
-	assert.Contains(t, buffer.String(), `'Go'`)
-	assert.Contains(t, buffer.String(), `'Java'`)
-	assert.Contains(t, buffer.String(), `'C'`)
-	assert.Contains(t, buffer.String(), `'C++'`)
-	assert.Contains(t, buffer.String(), `'Rust'`)
-	assert.Contains(t, buffer.String(), `'Closure'`)
-}
-
-func TestMySQLDumpTableDataHandlingErrorFromSelectAllDataFor(t *testing.T) {
-	db, mock := getDB(t)
-	buffer := bytes.NewBuffer(make([]byte, 0))
-	dumper := NewMySQLDumper(db)
-	error := errors.New("fail")
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnError(error)
-	assert.Equal(t, error, dumper.WriteTableData(buffer, "table"))
 }
